+++ jicek-license/src/main/java/com/jicek/license/enduser/mapper/EndUserMapper.java
@@ -1,0 +1,13 @@
+package com.jicek.license.enduser.mapper;
+
+import com.baomidou.mybatisplus.core.mapper.BaseMapper;
+import com.jicek.license.enduser.entity.EndUser;
+import org.apache.ibatis.annotations.Mapper;
+
+/**
+ * 终端用户 Mapper
+ * 作者: 极策k  日期: 2026-07-22
+ */
+@Mapper
+public interface EndUserMapper extends BaseMapper<EndUser> {
+}